requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.2
daphne==3.0.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.2.0
